nodejsが正しくインストールされない問題を解決

+++ run_once_install-packages.sh
@@ -4,8 +4,47 @@ function brew_install() {
     for N in "${@}";
     do
         if ! type $N > /dev/null 2>&1; then
+            brew_install_log ${N}
             brew install $N
         fi
+    done
+}
+
+function apt_install() {
+    for N in "${@}";
+    do
+        if ! type $N > /dev/null 2>&1; then
+            apt_install_log $N
+            sudo apt-get install -y $N
+        fi
+    done
+}
+
+function print_colored() {
+    ESCAPE=$(printf '\033') # \e や \x1bまたは$'\e'は使用しない
+    for N in "${@}";
+    do
+        printf "${ESCAPE}[31m%s ${ESCAPE}[m\n" $N
+    done
+}
+
+function apt_install_log() {
+    for N in "${@}";
+    do
+        print_colored "[apt install] $N"
+    done
+}
+function brew_install_log() {
+    for N in "${@}";
+    do
+        print_colored "[brew install] $N"
+    done
+}
+
+function install_log() {
+    for N in "${@}";
+    do
+        print_colored "[install] $N"
     done
 }
 
@@ -20,6 +59,7 @@ while [ $# -gt 0 ]; do
     shift
 done
 
+
 echo "is_all_install $is_all_install"
 
 sudo apt-get update
@@ -28,7 +68,7 @@ sudo apt-get -y autoremove
 
 # linuxbrew
 if ! type brew >/dev/null 2>&1; then
-    echo "[install] linuxbrew"
+    install_log linuxbrew
     /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
     test -d ~/.linuxbrew && eval "$(~/.linuxbrew/bin/brew shellenv)"
     test -d /home/linuxbrew/.linuxbrew && eval "$(/home/linuxbrew/.linuxbrew/bin/brew shellenv)"
@@ -39,10 +79,10 @@ fi
 brew update
 
 # python関係
-sudo apt-get install -y make build-essential libssl-dev zlib1g-dev \
-libbz2-dev libreadline-dev libsqlite3-dev wget curl llvm \
-libncursesw5-dev xz-utils tk-dev libxml2-dev libxmlsec1-dev libffi-dev liblzma-dev
 if ! type pyenv >/dev/null 2>&1; then
+    apt_install make build-essential libssl-dev zlib1g-dev \
+        libbz2-dev libreadline-dev libsqlite3-dev wget curl llvm \
+        libncursesw5-dev xz-utils tk-dev libxml2-dev libxmlsec1-dev libffi-dev liblzma-dev
     brew install pyenv
     pyenv install -f 3.8.10
     pyenv global 3.8.10
@@ -55,10 +95,11 @@ fi
 
 # nodejs
 if ! type npm >/dev/null 2>&1; then
-    sudo apt-get instal -y nodejs npm
+    apt_install nodejs npm
+    install_log "[npm install] npm"
     sudo npm install -g n
     sudo n stable
-    sudo apt-get remove -y --purge nodejs npm
+    sudo apt-get purge -y nodejs npm
 fi
 
 # git
@@ -71,21 +112,26 @@ brew_install starship exa bat
 
 # neovim環境構築
 if ! type nvim >/dev/null 2>&1; then
-    brew_install neovim, deno, fzf, xclip
+    brew_install neovim deno fzf xclip
 
     sudo npm install -g neovim
-    mkdir -p ~/python_envs/nvim/
-    cd ~/python_envs/nvim/
-    pipenv install pynvim
+    if ! [ -e '~/python_envs/nvim' ]; then
+        mkdir -p ~/python_envs/nvim/
+        cd ~/python_envs/nvim/
+        pipenv install pynvim
+    fi
 
-    curl https://raw.githubusercontent.com/Shougo/dein.vim/master/bin/installer.sh > ~/installer.sh
-    sh ~/installer.sh ~/.cache/dein
-    rm ~/installer.sh
+    if ! [ -e '~/.cache/dein' ]; then
+        curl https://raw.githubusercontent.com/Shougo/dein.vim/master/bin/installer.sh > ~/installer.sh
+        sh ~/installer.sh ~/.cache/dein
+        rm ~/installer.sh
+    fi
 fi
 
 
 # openMPI
 if [[ "$is_all_install" = true ]]; then
+    install_log openMPI
     sudo mkdir /opt/openMPI
     sudo rm /opt/openMPI/* -rf
     sudo apt-get install build-essential gfortran
@@ -100,6 +146,7 @@ fi
 
 # Git Credential Manager for Linux
 if ! type git-credential-manager-core >/dev/null 2>&1; then
+    install_log "git credential manager for linux"
     curl -LO https://raw.githubusercontent.com/GitCredentialManager/git-credential-manager/main/src/linux/Packaging.Linux/install-from-source.sh &&
     sh ./install-from-source.sh &&
     git-credential-manager-core configure
